tests/pytest_docker_compose_tests/docker_compose_directory/a_buildable_container/requirements.txt
uvicorn==0.7.1
psycopg2-binary==2.8.2
fastapi==0.20.0
